+++ inputenum.js
@@ -1,0 +1,9 @@
+let InputEnum = {
+	UP: "Up",
+	DOWN: "Down",
+	LEFT: "Left",
+	RIGHT: "Right"
+};
+
+Object.freeze(InputEnum);
+Object.seal(InputEnum);